+++ src/idc/generator/_list.py
@@ -45,6 +45,15 @@ class ListGenerator(SingleVariableGenerator):
         """
         return "Outputs the specified values."
 
+    def _default_var_name(self) -> str:
+        """
+        Returns the default variable name.
+
+        :return: the default name
+        :rtype: str
+        """
+        return "v"
+
     def _create_argparser(self) -> argparse.ArgumentParser:
         """
         Creates an argument parser. Derived classes need to fill in the options.