+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 ------------------
 
 - added generic plugins that take user Python functions: `from-pyfunc`, `pyfunc-filter`, `to-pyfunc`
+- added `idc-exec` tool that uses generator to produce variable/value pairs that are used to expand
+  the provided pipeline template which then gets executed
 
 
 0.0.1 (2024-05-06)